Download zip files to Downloads folder
This change modifies the download of a zip file that gets extracted
so that it ends up in the Downloads folder. This mimics more
closely the way it's working on TeamCity, and it allows to put all
files in a directory into an installer without the (unnecessary)
zip file laying around. Previously we had the zip file plus the
extracted files.