Handle missing 'View build configuration settings' and warn
In TeamCity after 7.0, the guest user no longer has permission to view build configuration settings.
This change will allow this to work with or without the permission, but allow the VCS configuration
to be displayed if the permission is given to the guest user (which has been done on build.palaso.org).